+++ src/Exceptions/AuthFailed.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace seregazhuk\PinterestBot\Exceptions;
-
-class AuthFailed extends PinterestBotException
-{
-
-}