+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Removed:
  - News *last* method
+ - AuthFailed exception
 
 ## [4.13.1] - 2016-10-15
 ### Fixed: